requirements.txt
Flask
Pillow
PyQt5
qrcode[pil]
streamlit
